Web: Changed adding a theme using a class, moved the picture to a component, removed it from Common.
