+++ packages/client/src/pages/VersionHistory/Section/Body/VersionBadge.js
@@ -1,7 +1,35 @@
 import React from "react";
+import styled, { css } from "styled-components";
 import Text from "@docspace/components/text";
 import { ColorTheme, ThemeType } from "@docspace/components/ColorTheme";
-import VersionMarkIcon from "@docspace/common/components/VersionMarkIcon";
+import VersionSvg from "PUBLIC_DIR/images/versionrevision_active.react.svg";
+
+const VersionMarkIcon = styled(VersionSvg)`
+  path {
+    fill: ${(props) =>
+      !props.$isVersion
+        ? props.theme.filesVersionHistory.badge.defaultFill
+        : props.index === 0
+        ? props.theme.filesVersionHistory.badge.fill
+        : props.theme.filesVersionHistory.badge.badgeFill};
+    stroke: ${(props) =>
+      !props.$isVersion
+        ? props.theme.filesVersionHistory.badge.stroke
+        : props.index === 0
+        ? props.theme.filesVersionHistory.badge.fill
+        : props.theme.filesVersionHistory.badge.badgeFill};
+
+    stroke-dasharray: ${(props) => (props.$isVersion ? "2 0" : "3 1")};
+    stroke-linejoin: ${(props) => (props.$isVersion ? "unset" : "round")};
+
+    ${(props) =>
+      props.$isVersion &&
+      css`
+        stroke-width: 2;
+      `}
+  }
+`;
+
 const VersionBadge = ({
   className,
   isVersion,
@@ -21,7 +49,12 @@ const VersionBadge = ({
     index={index}
     {...rest}
   >
-    <VersionMarkIcon $isVersion={isVersion} theme={theme} index={index} />
+    <VersionMarkIcon
+      className="version-mark-icon"
+      $isVersion={isVersion}
+      theme={theme}
+      index={index}
+    />
 
     <Text
       className="version_badge-text"
